+++ js/constants.js
@@ -1,0 +1,16 @@
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const NAMES = ['Милана', 'Владимир', 'Кирилл', 'Виктория', 'Вероника', 'Алиса', 'Каролина', 'Злата', 'Тимофей', 'Амелия', 'Иван', 'Мирослав', 'Валентина', 'Ян', 'Алина', 'Дмитрий', 'Максим', 'Адам', 'Арина', 'Алёна'];
+
+const PHOTOS_COUNT = 25;
+const COMMENTS_COUNT = 300;
+const AVATARS_COUNT = 6;
+
+export {COMMENTS, NAMES, PHOTOS_COUNT, COMMENTS_COUNT, AVATARS_COUNT};
